+++ app/(payload)/admin/importMap.js
@@ -1,6 +1,8 @@
 import { default as default_c8b6203be14e0a339519775c27f7b50f } from '@/components/admin/UsersCredentialsFormSync'
 import { default as default_cfe8833dd4642d6046713de140337991 } from '@/components/admin/UsersLocalPasswordFields'
 import { default as default_25d75f094cb6de9093023d79a509b859 } from '@/components/admin/UsersLoginMethodDisplay'
+import { default as default_3d5ce89c464db94aedb7076c07c0af9d } from '@/components/admin/HubspotSyncNowButton'
+import { default as default_ad497af92bb660355af21ff3a549e177 } from '@/components/admin/ApiKeyRowControls'
 import { default as default_9c986bf00998f72758fbf6c76782cb67 } from '@/components/admin/GoogleAdminLoginButton'
 import { default as default_9490f073a9408ebc42a7406874d5bb3f } from '@/components/admin/LocalAdminLoginView'
 import { CollectionCards as CollectionCards_f9c02e79a4aed9a3924487c0cd4cafb1 } from '@payloadcms/next/rsc'
@@ -10,6 +12,8 @@ export const importMap = {
   "@/components/admin/UsersCredentialsFormSync#default": default_c8b6203be14e0a339519775c27f7b50f,
   "@/components/admin/UsersLocalPasswordFields#default": default_cfe8833dd4642d6046713de140337991,
   "@/components/admin/UsersLoginMethodDisplay#default": default_25d75f094cb6de9093023d79a509b859,
+  "@/components/admin/HubspotSyncNowButton#default": default_3d5ce89c464db94aedb7076c07c0af9d,
+  "@/components/admin/ApiKeyRowControls#default": default_ad497af92bb660355af21ff3a549e177,
   "@/components/admin/GoogleAdminLoginButton#default": default_9c986bf00998f72758fbf6c76782cb67,
   "@/components/admin/LocalAdminLoginView#default": default_9490f073a9408ebc42a7406874d5bb3f,
   "@payloadcms/next/rsc#CollectionCards": CollectionCards_f9c02e79a4aed9a3924487c0cd4cafb1